+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+* Added placeholder for processing Grab&Go orders #APPS-1180
+
+### Updated
+* groue/GRDB.swift 6.23.0 (was 6.21.0)
+
 ## [0.43.2] - 2023-12-04
 
 ### Added